+++ network/fitness_calculator.cuh
@@ -10,7 +10,7 @@
 
 void
 calculateFitness(int *labels, float *images, float *networks, int networkCount, int dataCount,
-                 int *fitness, MODE mode);
+                 float *fitness, MODE mode);
 
 
 #endif //GA_CNN_FITNESS_CALCULATOR_CUH
